UI 좀 뚝딱뚝

+++ NWT-Dungeon/NWT-Dungeon/TitleScene.cpp
@@ -36,7 +36,7 @@ bool TitleScene::Title()
 MenuType TitleScene::MenuTypeRender()	
 {
 	COORD Resolution = GetConsoleResolution();
-	int x = Resolution.X / 2.2;
+	int x = (int)(Resolution.X / 2.2);
 	int y = Resolution.Y / 2;
 	int origin = y;
